Server is able to start and connect to database, Recipe class is giving errors about org.postgresql.util.PSQLException: ERROR: value too long for type character varying(255)

+++ Backend/demo/src/main/java/com/example/demo/FridgeItem/FridgeItem.java
@@ -17,7 +17,7 @@ public class FridgeItem {
     private int itemId;
 
     @Column(name = "quantity")
-    private int quantity;
+    private float quantity;
 
     public FridgeItem(int userId, int itemId, int quantity) {
         this.userId = userId;
@@ -33,7 +33,7 @@ public class FridgeItem {
         return itemId;
     }
 
-    public int getQuantity() {
+    public float getQuantity() {
         return quantity;
     }
 }
